[AOS] refactor : 2차 PR 피드백 적용 3번째 커밋 - RecyclerView 1개만 사용하도록 리팩토링 - 각 섹션마다 배경색이 달라지도록 확장함수 구현
