--- Android/app/src/main/java/com/codesquadhan/sidedish/data/model/be/MainResponse.kt
@@ -1,3 +1,0 @@
-package com.codesquadhan.sidedish.data.model.be
-
-class MainResponse: ArrayList<MainResponseItem>()
