+++ Android/app/src/main/java/com/codesquadhan/sidedish/network/MenuService.kt
@@ -1,8 +1,6 @@
 package com.codesquadhan.sidedish.network
 
-import com.codesquadhan.sidedish.data.model.be.MainResponse
 import com.codesquadhan.sidedish.data.model.be.MainResponseItem
-import com.codesquadhan.sidedish.data.model.onban.OnbanMainResponse
 import retrofit2.Response
 import retrofit2.http.GET
