+++ Android/app/src/main/java/com/codesquadhan/sidedish/ui/common/ViewType.kt
@@ -1,8 +1,8 @@
-package com.codesquadhan.sidedish.ui.constant
+package com.codesquadhan.sidedish.ui.common
 
 object ViewType {
 
     const val HEADER_VIEW_TYPE = 1
     const val FOOD_VIEW_TYPE = 2
-    //const val FOOD_VIEW_TYPE = 0
+
 }
