+++ Android/app/src/main/java/com/codesquadhan/sidedish/ui/common/bindRecylcerViewColor.kt
@@ -1,0 +1,14 @@
+package com.codesquadhan.sidedish.ui.common
+
+import android.view.View
+import androidx.core.content.ContextCompat
+import androidx.databinding.BindingAdapter
+import com.codesquadhan.sidedish.R
+
+@BindingAdapter("isWhite")
+fun bindRecyclerViewColor(view: View, isWhite: Boolean) {
+    when (isWhite) {
+        true -> view.setBackgroundColor(ContextCompat.getColor(view.context, R.color.White))
+        false -> view.setBackgroundColor(ContextCompat.getColor(view.context, R.color.Grey4))
+    }
+}
