+++ Android/app/src/main/java/com/codesquadhan/sidedish/ui/common/extensionfunction/addAllAndSetIsWhite.kt
@@ -1,0 +1,11 @@
+package com.codesquadhan.sidedish.ui.common
+
+import com.codesquadhan.sidedish.data.model.be.MainResponseItem
+
+fun <T> MutableList<T>.addAllAndSetIsWhite(menuList: List<T>) {
+    this.addAll(menuList)
+    menuList.forEach {
+        it as MainResponseItem
+        it.isWhite = false
+    }
+}
