+++ Android/app/src/main/java/com/codesquadhan/sidedish/ui/main/MainAdapter.kt
@@ -1,26 +1,19 @@
 package com.codesquadhan.sidedish.ui.main
 
-import android.content.Context
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.codesquadhan.sidedish.R
-import com.codesquadhan.sidedish.data.model.MenuData
 import com.codesquadhan.sidedish.data.model.be.MainResponseItem
 import com.codesquadhan.sidedish.databinding.ItemMainFoodBinding
 import com.codesquadhan.sidedish.databinding.ItemMainHeaderBinding
-import com.codesquadhan.sidedish.ui.constant.ViewType.FOOD_VIEW_TYPE
-import com.codesquadhan.sidedish.ui.constant.ViewType.HEADER_VIEW_TYPE
-import java.lang.RuntimeException
+import com.codesquadhan.sidedish.ui.common.ViewType.FOOD_VIEW_TYPE
+import com.codesquadhan.sidedish.ui.common.ViewType.HEADER_VIEW_TYPE
 
 class MainAdapter(private val itemClick: (id: Int) -> Unit) :
     ListAdapter<MainResponseItem, RecyclerView.ViewHolder>(diffUtil) {
-
-    private var isWhite = true
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return when (viewType) {
@@ -38,7 +31,6 @@ class MainAdapter(private val itemClick: (id: Int) -> Unit) :
                 MainFoodViewHolder(binding)
             }
             else -> {
-                Log.d("AppTest", "${viewType ?: -1}")
                 throw RuntimeException("Invalid ViewHolder Type")
             }
         }
@@ -62,21 +54,16 @@ class MainAdapter(private val itemClick: (id: Int) -> Unit) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(mainResponseItem: MainResponseItem) {
-            if (isWhite) binding.root.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.White))
-            else binding.root.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.Grey4))
-            isWhite = !isWhite
-
             binding.mainResponseItem = mainResponseItem
             binding.tvMainHeader.text = mainResponseItem.headerText
         }
+
     }
 
     inner class MainFoodViewHolder(private val binding: ItemMainFoodBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(mainResponseItem: MainResponseItem, itemClick: (id: Int) -> Unit) {
-            if (!isWhite) binding.root.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.White))
-            else binding.root.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.Grey4))
             binding.mainResponseItem = mainResponseItem
 
             binding.root.setOnClickListener {
