+++ Android/app/src/main/res/layout/item_main_food.xml
@@ -12,6 +12,7 @@
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        isWhite="@{mainResponseItem.isWhite}"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
